# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxury Hotel System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
...
  /rooms/{id}/status:
    put:
      operationId: setRoomStatus
      tags: [rooms]
      summary: Override a room status (e.g. cleaning, maintenance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomStatusUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
  /ota/channels/{id}/ingest:
    post:
      operationId: ingestOtaBooking
      tags: [ota]
      summary: Ingest a single booking received from an OTA channel (idempotent by externalRef)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtaIngestRequest"
      responses:
        "200":
          description: Booking ingested or already present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
components:
  schemas:
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        city:
          type: string
        address:
          type: string
        description:
          type: string
        rating:
          type: number
        imageUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
        priceFrom:
          type: number
        totalRooms:
          type: integer
        availableRooms:
          type: integer
        phone:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - location
        - city
        - address
        - description
        - rating
        - imageUrl
        - amenities
        - priceFrom
        - totalRooms
        - availableRooms
        - phone
        - email
...
    Room:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        roomNumber:
          type: string
        type:
          type: string
        description:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        imageUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        status:
          type: string
          enum: [available, reserved, occupied, cleaning, maintenance]
        floor:
          type: integer
        view:
          type: string
      required:
        - id
        - hotelId
        - roomNumber
        - type
        - description
        - pricePerNight
        - capacity
        - imageUrl
        - amenities
        - isAvailable
        - status
        - floor
        - view
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        hotelId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        specialRequests:
          type: string
        status:
          type: string
        source:
          type: string
        externalRef:
          type: string
          nullable: true
        checkedInAt:
          type: string
          nullable: true
        checkedOutAt:
          type: string
          nullable: true
        guestId:
          type: integer
          nullable: true
        totalPrice:
          type: number
        createdAt:
          type: string
        room:
          $ref: "#/components/schemas/Room"
        hotel:
          $ref: "#/components/schemas/Hotel"
      required:
        - id
        - roomId
        - hotelId
        - guestName
        - guestEmail
        - guestPhone
        - checkInDate
        - checkOutDate
        - numberOfGuests
        - status
        - source
        - totalPrice
        - createdAt
    RoomStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [available, reserved, occupied, cleaning, maintenance]
      required: [status]
...
    OtaIngestRequest:
      type: object
      properties:
        externalRef:
          type: string
          description: OTA's unique booking reference (used for idempotency)
        roomId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        totalPrice:
          type: number
        specialRequests:
          type: string
      required:
        - externalRef
        - roomId
        - guestName
        - guestEmail
        - guestPhone
        - checkInDate
        - checkOutDate
        - numberOfGuests
        - totalPrice