# lib/api-zod/src/generated/types/otaIngestRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */

export interface OtaIngestRequest {
  /** OTA's unique booking reference (used for idempotency) */
  externalRef: string;
  roomId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  totalPrice: number;
  specialRequests?: string;
}