# lib/api-zod/src/generated/types/roomStatusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatusUpdateStatus } from "./roomStatusUpdateStatus";

export interface RoomStatusUpdate {
  status: RoomStatusUpdateStatus;
}